Implement NoSQL (MongoDB) in MovieMicroService

+++ MovieMicroService/Models/Pagination/PageViewModel.cs
@@ -6,9 +6,9 @@ namespace MovieMicroService.Models.Pagination
     {
         public int PageNumber { get; set; }
         public int TotalPages { get; set; }
-        public int Count { get; set; }
+        public long Count { get; set; }
 
-        public PageViewModel(int count, int pageNumber, int pageSize)
+        public PageViewModel(long count, int pageNumber, int pageSize)
         {
             Count = count;
             PageNumber = pageNumber;
